1.管理端-设备，新增`所属用户`字段，只有超级用户可以修改该字段。该字段是管理端的注册用户。 2.管理端-设备，超级用户可以查看所有设备，普通用户只能查看属于自己的设备 3.init.py脚本可选是否删除旧数据库，可选是否创建超级用户

+++ init.py
@@ -2,10 +2,15 @@
 """Django's command-line utility for administrative tasks."""
 import os
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'django_scrcpy.settings')
-try:
-    os.remove('db.sqlite3')
-except Exception as e:
-    pass
+
+# 1.清空数据库
+if input("(1).是否清除旧数据库：1是，0否 ___ ") == '1':
+    try:
+        os.remove('db.sqlite3')
+        print("成功清除旧数据库信息")
+    except Exception as e:
+        pass
+print()
 try:
     from django.core.management import execute_from_command_line
 except ImportError as exc:
@@ -14,9 +19,18 @@ except ImportError as exc:
         "available on your PYTHONPATH environment variable? Did you "
         "forget to activate a virtual environment?"
     ) from exc
+
+# 2.更新数据库结构和静态文件
+print("(2).更新数据库结构和静态文件")
 execute_from_command_line(['manage.py', 'migrate'])
 execute_from_command_line(['manage.py', 'collectstatic', '--noinput'])
-print('\n==================================================')
-print('Please create a super user !!!')
-print('==================================================')
-execute_from_command_line(['manage.py', 'createsuperuser'])
+print("更新成功")
+print()
+
+# 3.创建超级用户
+if input("(3).是否创建超级用户：1是，0否 ___ ") == '1':
+    print('\n==================================================')
+    print('Please create a super user !!!')
+    print('==================================================')
+    execute_from_command_line(['manage.py', 'createsuperuser'])
+    print("创建成功")
